Uptil passport-JWT

+++ controllers/api/v1/posts_api.js
@@ -21,23 +21,25 @@ module.exports.index = async (req, res) => {
 module.exports.destroy = async (req, res) => {
   try {
     let post = await Post.findById(req.params.id);
-    // if (post.user == req.user.id) {
-    await post.remove();
-    await Comment.deleteMany({ post: req.params.id });
+    if (post.user == req.user.id) {
+      await post.remove();
+      await Comment.deleteMany({ post: req.params.id });
 
-    // if (req.xhr) {
-    //   return res.status(200).json({
-    //     data: {
-    //       post_id: req.params.id,
-    //     },
-    //     message: "Post deleted",
-    //   });
-    // }
+      // if (req.xhr) {
+      //   return res.status(200).json({
+      //     data: {
+      //       post_id: req.params.id,
+      //     },
+      //     message: "Post deleted",
+      //   });
+      // }
 
-    // req.flash("success", "Post and associated comments deleted!");
-    // } else {
-    // req.flash("error", "You cannot delete this post!");
-    // }
+      req.flash("success", "Post and associated comments deleted!");
+    } else {
+      return res.status(401).json({
+        message: "You cannot delete this post",
+      });
+    }
     return res.status(200).json({
       message: "Post and associated comments deleted successfully",
     });
